+ Created BuildViewRotationMatrix and BuildTranslationMatrix methods for MVPMatrix. We can now create a view matrix that depends on a "camera's" position and orientation. View rotation and model rotation matrices are different, however. View's matrix is has to be fixed, with regards to what is considered "forward" for it.
This is because in OpenGL, the default view is down the z axis (positive z is behind the viewer). So Forward is (0,0,-1) and Up is (0,1,0). In order to turn this into the default view matrix, the forward vector is flipped and taken as z and x is built from the cross product of (Up * z), and results in (1,0,0). This allows us to create the identity matrix using x, up and z as the first three columns, with a 1 in the last element of the fourth column and zeros everywhere else.